+++ tuplities-flatten-nest/src/lib.rs
@@ -8,7 +8,9 @@ mod nest;
 mod nested_index;
 mod nested_option;
 mod nested_push_pop;
+mod nested_ref;
 mod nested_replicate;
+mod nested_row;
 mod nested_try_from;
 
 // Re-export all public traits and types
@@ -22,5 +24,7 @@ pub use nested_option::{IntoNestedTupleOption, NestedTupleOption};
 pub use nested_push_pop::{
     NestedTuplePopBack, NestedTuplePopFront, NestedTuplePushBack, NestedTuplePushFront,
 };
+pub use nested_ref::{NestedTupleMut, NestedTupleRef};
 pub use nested_replicate::NestedTupleReplicate;
+pub use nested_row::{NestedTupleRow, NestedTupleRowMut};
 pub use nested_try_from::{NestedTupleTryFrom, NestedTupleTryInto};