+++ tuplities/src/lib.rs
@@ -17,8 +17,9 @@ pub mod prelude {
     pub use tuplities_flatten_nest::{
         FlattenMatrixElements, FlattenNestedTuple, FlattenNestedTupleMatrix, IntoNestedTupleOption,
         NestMatrixElements, NestTuple, NestTupleMatrix, NestedTupleIndex, NestedTupleIndexMut,
-        NestedTupleOption, NestedTuplePopBack, NestedTuplePopFront, NestedTuplePushBack,
-        NestedTuplePushFront, NestedTupleReplicate, NestedTupleTryFrom, NestedTupleTryInto,
+        NestedTupleMut, NestedTupleOption, NestedTuplePopBack, NestedTuplePopFront,
+        NestedTuplePushBack, NestedTuplePushFront, NestedTupleRef, NestedTupleReplicate,
+        NestedTupleRow, NestedTupleRowMut, NestedTupleTryFrom, NestedTupleTryInto,
     };
     #[cfg(feature = "from")]
     pub use tuplities_from::{TupleFrom, TupleInto};